Consolidate the config schemas
