--- lib/cloud_controller/config_schemas/vms/migrate_schema.rb
@@ -1,10 +1,0 @@
-require 'vcap/config'
-require 'cloud_controller/config_schemas/base/migrate_schema'
-
-module VCAP::CloudController
-  module ConfigSchemas
-    module Vms
-      MigrateSchema = VCAP::CloudController::ConfigSchemas::Base::MigrateSchema
-    end
-  end
-end
